fix: move menu into layouts and layouts folder should be inside components
